% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_pkg.R
\name{bootstrap_pkg}
\alias{bootstrap_pkg}
\title{Create an opinionated R package}
\usage{
bootstrap_pkg(
  path,
  author_name = "Bjarke Hautop Kristensen",
  author_email = "bjarke.hautop@gmail.com",
  license = usethis::use_mit_license(),
  private = FALSE
)
}
\arguments{
\item{path}{Path where the new package is created. The last component of
the path is used as the package name.}

\item{author_name}{Name of the package author (aut/cre in \code{Authors@R} and
the license copyright holder).}

\item{author_email}{Email of the package author.}

\item{license}{A call to a usethis license function, e.g.
\code{usethis::use_mit_license()} or \code{usethis::use_gpl3_license()}.}

\item{private}{Should the GitHub repository be private?}
}
\value{
The path to the new package, invisibly.
}
\description{
Creates a new package with \code{\link[usethis:create_package]{usethis::create_package()}} and applies an
opinionated setup:
}
\details{
\itemize{
\item air formatting (\code{\link[usethis:use_air]{usethis::use_air()}})
\item git, pushed to GitHub (\code{\link[usethis:use_git]{usethis::use_git()}}, \code{\link[usethis:use_github]{usethis::use_github()}})
\item a license, MIT by default
\item an Rmd README (\code{\link[usethis:use_readme_rmd]{usethis::use_readme_rmd()}})
\item codecov test coverage with an opinionated \code{codecov.yml}
(\code{\link[usethis:use_coverage]{usethis::use_coverage()}})
\item spell checking (\code{\link[usethis:use_spell_check]{usethis::use_spell_check()}})
\item \code{R-CMD-check} and \code{test-coverage} GitHub Actions workflows
(\code{\link[usethis:use_github_action]{usethis::use_github_action()}})
\item Dependabot updates for GitHub Actions
\item a pkgdown site deployed via GitHub Pages
(\code{\link[usethis:use_pkgdown_github_pages]{usethis::use_pkgdown_github_pages()}})
\item a \code{.pre-commit-config.yaml} with panache and air hooks pinned to their
latest release, installed with \href{https://github.com/j178/prek}{prek} if
available
}
}
